refactor(P4-2): Server Action ファイルを統一（Batch2）
- 単数 action.ts 3件を actions.ts へ（schedule / schedule-generator / ganttv3、
  相対 import 含め全更新）
- app/actions/ 配下を機能セグメントへ移動:
  - notification-actions.ts → app/notifications/
  - evm/evm-actions.ts → app/wbs/[id]/actions/
  - 空になった app/actions/ を解消

本番型チェッククリーン。
