+++ src/components/evm/evm-dashboard.tsx
@@ -14,7 +14,7 @@ import { EvmChart } from "./evm-chart";
 import { EvmMetricsCard } from "./evm-metrics-card";
 import { TaskEvmTable } from "./task-evm-table";
 import { EvmTimeSeriesTable } from "./evm-timeseries-table";
-import { getEvmDashboardData } from "@/app/actions/evm/evm-actions";
+import { getEvmDashboardData } from "@/app/wbs/[id]/actions/evm-actions";
 import type { EvmMetricsData, TaskEvmDataSerialized } from "@/applications/evm/evm-dashboard-dto";
 import { Loader2, TrendingUp, DollarSign, Info } from "lucide-react";
 import { Alert, AlertDescription } from "@/components/ui/alert";
